Production-grade Rebranding and Stabilization: Bharat Dues
- Migrated entire financial engine to double-precision (2-decimal) logic.
- Implemented epsilon-based rounding for perfect balance matching.
- Rebranded 'SettleUp Lite' to 'Bharat Dues' across UI, Android, and iOS.
- Fixed UI overflows in payment confirmation and expense modals.
- Enhanced activity logs with granular field auditing.
- Fully configured Web support with managed Google Client IDs.
- Restored stable Google Sign-In implementation (6.2.1).
- Added multi-platform (iOS/macOS) URL schemes and permissions.
- Updated documentation and project architecture overview.

+++ lib/models/expense.dart
@@ -58,9 +58,9 @@ class Expense {
 
   final String id;
   final String description;
-  final int amount;
+  final double amount;
   final String payerId;
-  final Map<String, int> shares;
+  final Map<String, double> shares;
   final DateTime createdAt;
   final String createdBy;
   final ExpenseCategory? category;
@@ -84,7 +84,7 @@ class Expense {
   factory Expense.fromJson(String id, Map<String, dynamic> json) {
     final sharesRaw = json['shares'] as Map<String, dynamic>? ?? {};
     final shares = sharesRaw.map(
-      (key, value) => MapEntry(key, (value as num).toInt()),
+      (key, value) => MapEntry(key, (value as num).toDouble()),
     );
 
     DateTime createdAt;
@@ -106,7 +106,7 @@ class Expense {
     return Expense(
       id: id,
       description: json['description'] as String? ?? 'Expense',
-      amount: (json['amount'] as num?)?.toInt() ?? 0,
+      amount: (json['amount'] as num?)?.toDouble() ?? 0.0,
       payerId: json['payerId'] as String? ?? '',
       shares: shares,
       createdAt: createdAt,
